addes DataGenerator
