--- DataGenerator/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.example</groupId>
-    <artifactId>DataGenerator</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <dependencies>
-
-        <!-- https://mvnrepository.com/artifact/com.devskiller/jfairy -->
-        <dependency>
-            <groupId>com.devskiller</groupId>
-            <artifactId>jfairy</artifactId>
-            <version>0.6.4</version>
-        </dependency>
-
-
-    </dependencies>
-
-</project>
